requirements.txt
tensorflow
numpy
pillow
matplotlib
scikit-learn
seaborn
opencv-python
streamlit